requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
tensorflow==2.15.0
tf2onnx==1.16.0
onnx==1.15.0
onnxruntime==1.17.0
cryptography==41.0.7
pytest==8.0.0
python-dotenv==1.0.1
aiofiles==23.2.1
bcrypt==4.1.2
